+++ R/zzz.R
@@ -1,4 +1,5 @@
 .onLoad <- function(libname, pkgname) {
+  set_echo(confirm = FALSE)
   set_log_level(confirm = FALSE, level = 3L)
   set_logfile(confirm = FALSE, create = FALSE)
   set_timestamp_format(confirm = FALSE)