+++ NEWS.md
@@ -14,6 +14,7 @@
   This was previously (unintentionally) guaranteed by replacing the `NA` with `"__NA__"`.
 * A global log level can now be set using `set_log_level()`, which is used by all functions unless otherwise stated.
   The log levels are: `"DEBUG"`, `"INFO"`, `"WARN"`, `"ERROR"` and `"NONE"`.
+* Add `set_echo()` to control globally whether log messages are echoed to the console.
 
 ## Bugfixes 
 * `read_logs()` now correctly reads empty character values `""`, as in `{"key": ""}`, as such.